Refactor API

+++ src/main/java/com/example/exchange/rate/service/modal/ExchangeRates.java
@@ -3,11 +3,12 @@ package com.example.exchange.rate.service.modal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 @AllArgsConstructor
 public class ExchangeRates {
-    private Map<String, Double> exchangeRates;
+    private Map<String, BigDecimal> exchangeRates;
     private String date;
 }
